Define memory resources in gigabytes
The ‘resource’ part of a snakemake rule can only be defined with integers. Gridengine memory definitions usually have a suffice K,M,G for kilo, mega, gigabyte. With no suffix the number is interpreted as bytes.

We now check fo memory resource requests and append ‘G’ to the end of the integer, thus interpreting all memory requests in gigabytes.