Fix cluster.yaml not being used
We were checking if a set of options is a dict and if not then we return (silently failing). The yaml file is actually read into an OrderedDict so this was not being processed. I’ve removed the check because it’s better that we fail than silently fail. The default cluster.yaml is now read.