Apply a list of non-requestable resources
Depending on the local cluster configuration some resources might not be user requestable. These can now be explicitly listed.

Most importantly snakemake from version 7 onwards now gives a default tmpdir resource, but this is not requestable usually in gridengine, so we remove this. In the future it would be sensible to use this to set the TMPDIR environment variable.