Support snakemake mem_mb and disk_mb
From snakemake 7 'mem_mb' and 'disk_mb' are default resources which are always passed in. We map these to h_vmem and h_fsize.

Note that now we change the units of memory to be in megabytes (previously we were using (gigabytes). This will be a breaking change for old snakemake rules.