Add support for ignoring resources
Snakemake now defines some default resources, not all of which are supported by grid engine, e.g. memory and disk requests in mebibytes (mem_mib, disk_mib).

This patch allows a list of resources to be hardcoded to be ignored by gridengine. There is probably a more elgant way to do this, but it basically works.

I also considered simply stripping out resources which are not available in the cluster definition, but this could lead to silent errors when the list of resources in misconfigured.